Strip spaces from private key
The private key file sometimes contains a newline. We should strip
spaces so those files are interpreted correctly.

I also changed it so EncryptionManager only accepts hex-encoded keys,
since it only returns hex-encoded keys from new_keypair as well. Not
sure what that was used for (doesn't seem to be a test for it).