Improve readme and add warning on rails credentials conflict (#35)
Eyaml doesn't work with the new rails' credentials encryption (they both do the same thing in principle). The rails credentials management relies on a .yml.encfile which is encrypted via a key present in the master.key file (which is 32 bytes). Eyaml uses it own key which is 64 bytes in fact. I've improved the readme for general eyaml config/installation and added a raise when a master/key file is present.